SCRUM-23 feat: solution ok
